Move platform extensions to separate file
